# contracts/openapi/pipe4-v1.json
{
  "components": {
    "schemas": {
...
      "ApiErrorBody": {
        "additionalProperties": false,
        "properties": {
          "code": {
            "title": "Code",
            "type": "string"
          },
          "details": {
            "anyOf": [
              {
                "additionalProperties": true,
                "type": "object"
              },
              {
                "type": "null"
              }
            ],
            "title": "Details"
          },
          "message": {
            "title": "Message",
            "type": "string"
          }
        },
        "required": [
          "code",
          "message"
        ],
        "title": "ApiErrorBody",
        "type": "object"
      },
      "ApiFailure": {
        "additionalProperties": false,
        "properties": {
          "error": {
            "$ref": "#/components/schemas/ApiErrorBody"
          },
          "meta": {
            "$ref": "#/components/schemas/ApiMeta"
          },
          "success": {
            "default": false,
            "title": "Success",
            "type": "boolean"
          }
        },
        "required": [
          "error",
          "meta"
        ],
        "title": "ApiFailure",
        "type": "object"
      },
      "ApiMeta": {
        "additionalProperties": false,
        "properties": {
          "request_id": {
            "title": "Request Id",
            "type": "string"
          },
          "timestamp": {
            "format": "date-time",
            "title": "Timestamp",
            "type": "string"
          }
        },
        "required": [
          "request_id",
          "timestamp"
        ],
        "title": "ApiMeta",
        "type": "object"
      },
...
      "ApiSuccess_CurrentStatePresentation_": {
        "additionalProperties": false,
        "properties": {
          "data": {
            "$ref": "#/components/schemas/CurrentStatePresentation"
          },
          "meta": {
            "$ref": "#/components/schemas/ApiMeta"
          },
          "success": {
            "default": true,
            "title": "Success",
            "type": "boolean"
          }
        },
        "required": [
          "data",
          "meta"
        ],
        "title": "ApiSuccess[CurrentStatePresentation]",
        "type": "object"
      },
...
      "ApiSuccess_ObservationRecord_": {
        "additionalProperties": false,
        "properties": {
          "data": {
            "$ref": "#/components/schemas/ObservationRecord"
          },
          "meta": {
            "$ref": "#/components/schemas/ApiMeta"
          },
          "success": {
            "default": true,
            "title": "Success",
            "type": "boolean"
          }
        },
        "required": [
          "data",
          "meta"
        ],
        "title": "ApiSuccess[ObservationRecord]",
        "type": "object"
      },
...
      "BooleanStateValue": {
        "additionalProperties": false,
        "properties": {
          "kind": {
            "const": "boolean",
            "default": "boolean",
            "title": "Kind",
            "type": "string"
          },
          "value": {
            "title": "Value",
            "type": "boolean"
          }
        },
        "required": [
          "value"
        ],
        "title": "BooleanStateValue",
        "type": "object"
      },
      "CategoricalStateValue": {
        "additionalProperties": false,
        "properties": {
          "kind": {
            "const": "categorical",
            "default": "categorical",
            "title": "Kind",
            "type": "string"
          },
          "value": {
            "maxLength": 120,
            "minLength": 1,
            "title": "Value",
            "type": "string"
          }
        },
        "required": [
          "value"
        ],
        "title": "CategoricalStateValue",
        "type": "object"
      },
      "ChangeStateValue": {
        "additionalProperties": false,
        "properties": {
          "category": {
            "anyOf": [
              {
                "maxLength": 80,
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Category"
          },
          "changed": {
            "title": "Changed",
            "type": "boolean"
          },
          "kind": {
            "const": "change",
            "default": "change",
            "title": "Kind",
            "type": "string"
          },
          "previous_value_hash": {
            "anyOf": [
              {
                "maxLength": 128,
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Previous Value Hash"
          },
          "summary": {
            "anyOf": [
              {
                "maxLength": 500,
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Summary"
          }
        },
        "required": [
          "changed"
        ],
        "title": "ChangeStateValue",
        "type": "object"
      },
...
      "CurrentStatePresentation": {
        "additionalProperties": false,
        "properties": {
          "access_request_status": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Access Request Status"
          },
          "actions": {
            "items": {
              "type": "string"
            },
            "title": "Actions",
            "type": "array"
          },
          "confidence": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Confidence"
          },
          "entity": {
            "$ref": "#/components/schemas/EntitySummary"
          },
          "epistemic_status": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/EpistemicStatus"
              },
              {
                "type": "null"
              }
            ]
          },
          "expires_at": {
            "anyOf": [
              {
                "format": "date-time",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Expires At"
          },
          "freshness_seconds": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Freshness Seconds"
          },
          "identity": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/StateIdentity"
              },
              {
                "type": "null"
              }
            ]
          },
          "kind": {
            "$ref": "#/components/schemas/PresentationKind"
          },
          "message": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Message"
          },
          "observed_at": {
            "anyOf": [
              {
                "format": "date-time",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Observed At"
          },
          "provenance": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/ProvenanceSummary"
              },
              {
                "type": "null"
              }
            ]
          },
          "state_type": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/StateType"
              },
              {
                "type": "null"
              }
            ]
          },
          "state_version_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "State Version Id"
          },
          "status": {
            "$ref": "#/components/schemas/UserFacingStateStatus"
          },
          "value": {
            "anyOf": [
              {
                "discriminator": {
                  "mapping": {
                    "boolean": "#/components/schemas/BooleanStateValue",
                    "categorical": "#/components/schemas/CategoricalStateValue",
                    "change": "#/components/schemas/ChangeStateValue",
                    "duration": "#/components/schemas/DurationStateValue",
                    "quantity": "#/components/schemas/QuantityStateValue",
                    "range": "#/components/schemas/RangeStateValue"
                  },
                  "propertyName": "kind"
                },
                "oneOf": [
                  {
                    "$ref": "#/components/schemas/BooleanStateValue"
                  },
                  {
                    "$ref": "#/components/schemas/QuantityStateValue"
                  },
                  {
                    "$ref": "#/components/schemas/DurationStateValue"
                  },
                  {
                    "$ref": "#/components/schemas/RangeStateValue"
                  },
                  {
                    "$ref": "#/components/schemas/ChangeStateValue"
                  },
                  {
                    "$ref": "#/components/schemas/CategoricalStateValue"
                  }
                ]
              },
              {
                "type": "null"
              }
            ],
            "title": "Value"
          },
          "verification_status": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/VerificationStatus"
              },
              {
                "type": "null"
              }
            ]
          }
        },
        "required": [
          "kind",
          "status",
          "entity"
        ],
        "title": "CurrentStatePresentation",
        "type": "object"
      },
...
      "DeclarationRequest": {
        "additionalProperties": false,
        "properties": {
          "idempotency_key": {
            "anyOf": [
              {
                "maxLength": 200,
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Idempotency Key"
          },
          "identity": {
            "$ref": "#/components/schemas/IdentityBody"
          },
          "observed_at": {
            "format": "date-time",
            "title": "Observed At",
            "type": "string"
          },
          "permission_tags": {
            "items": {
              "type": "string"
            },
            "title": "Permission Tags",
            "type": "array",
            "uniqueItems": true
          },
          "source_id": {
            "title": "Source Id",
            "type": "string"
          },
          "value": {
            "discriminator": {
              "mapping": {
                "boolean": "#/components/schemas/BooleanStateValue",
                "categorical": "#/components/schemas/CategoricalStateValue",
                "change": "#/components/schemas/ChangeStateValue",
                "duration": "#/components/schemas/DurationStateValue",
                "quantity": "#/components/schemas/QuantityStateValue",
                "range": "#/components/schemas/RangeStateValue"
              },
              "propertyName": "kind"
            },
            "oneOf": [
              {
                "$ref": "#/components/schemas/BooleanStateValue"
              },
              {
                "$ref": "#/components/schemas/QuantityStateValue"
              },
              {
                "$ref": "#/components/schemas/DurationStateValue"
              },
              {
                "$ref": "#/components/schemas/RangeStateValue"
              },
              {
                "$ref": "#/components/schemas/ChangeStateValue"
              },
              {
                "$ref": "#/components/schemas/CategoricalStateValue"
              }
            ],
            "title": "Value"
          },
          "visibility": {
            "$ref": "#/components/schemas/VisibilityScope",
            "default": "PUBLIC"
          }
        },
        "required": [
          "identity",
          "value",
          "observed_at",
          "source_id"
        ],
        "title": "DeclarationRequest",
        "type": "object"
      },
...
      "DurationStateValue": {
        "additionalProperties": false,
        "properties": {
          "kind": {
            "const": "duration",
            "default": "duration",
            "title": "Kind",
            "type": "string"
          },
          "label": {
            "anyOf": [
              {
                "maxLength": 120,
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Label"
          },
          "seconds": {
            "minimum": 0.0,
            "title": "Seconds",
            "type": "integer"
          }
        },
        "required": [
          "seconds"
        ],
        "title": "DurationStateValue",
        "type": "object"
      },
...
      "EntitySummary": {
        "additionalProperties": false,
        "properties": {
          "display_name": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Display Name"
          },
          "distance_m": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Distance M"
          },
          "entity_id": {
            "title": "Entity Id",
            "type": "string"
          },
          "entity_type": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Entity Type"
          }
        },
        "required": [
          "entity_id"
        ],
        "title": "EntitySummary",
        "type": "object"
      },
...
      "IdentityBody": {
        "additionalProperties": false,
        "properties": {
          "entity_id": {
            "title": "Entity Id",
            "type": "string"
          },
          "location_key": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Location Key"
          },
          "object_key": {
            "default": "main",
            "title": "Object Key",
            "type": "string"
          },
          "qualifiers": {
            "additionalProperties": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                },
                {
                  "type": "number"
                },
                {
                  "type": "boolean"
                }
              ]
            },
            "title": "Qualifiers",
            "type": "object"
          },
          "state_type": {
            "$ref": "#/components/schemas/StateType"
          }
        },
        "required": [
          "entity_id",
          "state_type"
        ],
        "title": "IdentityBody",
        "type": "object"
      },
...
      "ObservationRecord": {
        "additionalProperties": false,
        "properties": {
          "actor_subject": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Actor Subject"
          },
          "consumed_state_version_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Consumed State Version Id"
          },
          "correlation_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Correlation Id"
          },
          "epistemic_status": {
            "$ref": "#/components/schemas/EpistemicStatus",
            "default": "OBSERVED"
          },
          "evidence_ids": {
            "items": {
              "type": "string"
            },
            "title": "Evidence Ids",
            "type": "array"
          },
          "identity": {
            "$ref": "#/components/schemas/StateIdentity"
          },
          "interpretation_confidence": {
            "anyOf": [
              {
                "maximum": 1.0,
                "minimum": 0.0,
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Interpretation Confidence"
          },
          "latitude": {
            "anyOf": [
              {
                "maximum": 90.0,
                "minimum": -90.0,
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Latitude"
          },
          "location_accuracy_m": {
            "anyOf": [
              {
                "minimum": 0.0,
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Location Accuracy M"
          },
          "location_description": {
            "anyOf": [
              {
                "maxLength": 500,
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Location Description"
          },
          "location_evidence_type": {
            "$ref": "#/components/schemas/LocationEvidenceType",
            "default": "UNKNOWN"
          },
          "longitude": {
            "anyOf": [
              {
                "maximum": 180.0,
                "minimum": -180.0,
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Longitude"
          },
          "observation_id": {
            "title": "Observation Id",
            "type": "string"
          },
          "observed_at": {
            "format": "date-time",
            "title": "Observed At",
            "type": "string"
          },
          "original_input_ref": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Original Input Ref"
          },
          "proposed_value": {
            "discriminator": {
              "mapping": {
                "boolean": "#/components/schemas/BooleanStateValue",
                "categorical": "#/components/schemas/CategoricalStateValue",
                "change": "#/components/schemas/ChangeStateValue",
                "duration": "#/components/schemas/DurationStateValue",
                "quantity": "#/components/schemas/QuantityStateValue",
                "range": "#/components/schemas/RangeStateValue"
              },
              "propertyName": "kind"
            },
            "oneOf": [
              {
                "$ref": "#/components/schemas/BooleanStateValue"
              },
              {
                "$ref": "#/components/schemas/QuantityStateValue"
              },
              {
                "$ref": "#/components/schemas/DurationStateValue"
              },
              {
                "$ref": "#/components/schemas/RangeStateValue"
              },
              {
                "$ref": "#/components/schemas/ChangeStateValue"
              },
              {
                "$ref": "#/components/schemas/CategoricalStateValue"
              }
            ],
            "title": "Proposed Value"
          },
          "received_at": {
            "format": "date-time",
            "title": "Received At",
            "type": "string"
          },
          "source_id": {
            "title": "Source Id",
            "type": "string"
          },
          "status": {
            "$ref": "#/components/schemas/ObservationStatus",
            "default": "RECEIVED"
          }
        },
        "required": [
          "identity",
          "proposed_value",
          "source_id",
          "observed_at",
          "received_at"
        ],
        "title": "ObservationRecord",
        "type": "object"
      },
      "ObservationRequest": {
        "additionalProperties": false,
        "properties": {
          "identity": {
            "$ref": "#/components/schemas/IdentityBody"
          },
          "latitude": {
            "anyOf": [
              {
                "maximum": 90.0,
                "minimum": -90.0,
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Latitude"
          },
          "location_accuracy_m": {
            "anyOf": [
              {
                "minimum": 0.0,
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Location Accuracy M"
          },
          "location_description": {
            "anyOf": [
              {
                "maxLength": 500,
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Location Description"
          },
          "longitude": {
            "anyOf": [
              {
                "maximum": 180.0,
                "minimum": -180.0,
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Longitude"
          },
          "observed_at": {
            "format": "date-time",
            "title": "Observed At",
            "type": "string"
          },
          "source_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Source Id"
          },
          "value": {
            "discriminator": {
              "mapping": {
                "boolean": "#/components/schemas/BooleanStateValue",
                "categorical": "#/components/schemas/CategoricalStateValue",
                "change": "#/components/schemas/ChangeStateValue",
                "duration": "#/components/schemas/DurationStateValue",
                "quantity": "#/components/schemas/QuantityStateValue",
                "range": "#/components/schemas/RangeStateValue"
              },
              "propertyName": "kind"
            },
            "oneOf": [
              {
                "$ref": "#/components/schemas/BooleanStateValue"
              },
              {
                "$ref": "#/components/schemas/QuantityStateValue"
              },
              {
                "$ref": "#/components/schemas/DurationStateValue"
              },
              {
                "$ref": "#/components/schemas/RangeStateValue"
              },
              {
                "$ref": "#/components/schemas/ChangeStateValue"
              },
              {
                "$ref": "#/components/schemas/CategoricalStateValue"
              }
            ],
            "title": "Value"
          }
        },
        "required": [
          "identity",
          "value",
          "observed_at"
        ],
        "title": "ObservationRequest",
        "type": "object"
      },
...
      "QuantityStateValue": {
        "additionalProperties": false,
        "properties": {
          "available": {
            "title": "Available",
            "type": "number"
          },
          "kind": {
            "const": "quantity",
            "default": "quantity",
            "title": "Kind",
            "type": "string"
          },
          "total_capacity": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Total Capacity"
          },
          "unit": {
            "maxLength": 32,
            "minLength": 1,
            "title": "Unit",
            "type": "string"
          }
        },
        "required": [
          "available",
          "unit"
        ],
        "title": "QuantityStateValue",
        "type": "object"
      },
      "RangeStateValue": {
        "additionalProperties": false,
        "properties": {
          "kind": {
            "const": "range",
            "default": "range",
            "title": "Kind",
            "type": "string"
          },
          "maximum": {
            "title": "Maximum",
            "type": "number"
          },
          "minimum": {
            "title": "Minimum",
            "type": "number"
          },
          "unit": {
            "maxLength": 32,
            "minLength": 1,
            "title": "Unit",
            "type": "string"
          }
        },
        "required": [
          "minimum",
          "maximum",
          "unit"
        ],
        "title": "RangeStateValue",
        "type": "object"
      },
...
      "StateType": {
        "enum": [
          "AVAILABLE",
          "ACCESSIBLE",
          "WORKING",
          "TIME",
          "CHANGED"
        ],
        "title": "StateType",
        "type": "string"
      },
...
      "VisibilityScope": {
        "enum": [
          "PUBLIC",
          "TARGETED",
          "PRIVATE",
          "INSTITUTIONAL",
          "INTERNAL"
        ],
        "title": "VisibilityScope",
        "type": "string"
      },
...
    },
    "securitySchemes": {
      "APIKeyHeader": {
        "in": "header",
        "name": "X-Internal-Token",
        "type": "apiKey"
      },
      "HTTPBearer": {
        "scheme": "bearer",
        "type": "http"
      }
    }
  },
  "info": {
    "description": "Horizontal current-reality resolution, evidence, verification, WATCH, access, nearby and acquisition service.",
    "title": "UWACI Pipe 4 Current Reality API",
    "version": "1.0.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/api/v1/pipe4/declarations": {
      "post": {
        "operationId": "declare_api_v1_pipe4_declarations_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/DeclarationRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiSuccess_CurrentStatePresentation_"
                }
              }
            },
            "description": "Successful Response"
          },
          "400": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiFailure"
                }
              }
            },
            "description": "Invalid request intent"
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiFailure"
                }
              }
            },
            "description": "Authentication required"
          },
          "403": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiFailure"
                }
              }
            },
            "description": "Permission denied"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiFailure"
                }
              }
            },
            "description": "Resource not found"
          },
          "409": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiFailure"
                }
              }
            },
            "description": "State/resource conflict"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiFailure"
                }
              }
            },
            "description": "Validation failed"
          },
          "429": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiFailure"
                }
              }
            },
            "description": "Rate limited"
          },
          "500": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiFailure"
                }
              }
            },
            "description": "Unexpected server error"
          },
          "503": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiFailure"
                }
              }
            },
            "description": "Dependency unavailable"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Declare current state",
        "tags": [
          "Pipe 4 evidence"
        ]
      }
    },
...
    "/api/v1/pipe4/observations": {
      "post": {
        "operationId": "observe_api_v1_pipe4_observations_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ObservationRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "202": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiSuccess_ObservationRecord_"
                }
              }
            },
            "description": "Successful Response"
          },
          "400": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiFailure"
                }
              }
            },
            "description": "Invalid request intent"
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiFailure"
                }
              }
            },
            "description": "Authentication required"
          },
          "403": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiFailure"
                }
              }
            },
            "description": "Permission denied"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiFailure"
                }
              }
            },
            "description": "Resource not found"
          },
          "409": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiFailure"
                }
              }
            },
            "description": "State/resource conflict"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiFailure"
                }
              }
            },
            "description": "Validation failed"
          },
          "429": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiFailure"
                }
              }
            },
            "description": "Rate limited"
          },
          "500": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiFailure"
                }
              }
            },
            "description": "Unexpected server error"
          },
          "503": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiFailure"
                }
              }
            },
            "description": "Dependency unavailable"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Submit structured observation",
        "tags": [
          "Pipe 4 evidence"
        ]
      }
    },
...
  }
}